+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 - Fix `stats` adding it's self size twice, incorrectly increasing the MB size.
+- Fixed paths being always joined as `some/location`, now Deno's `join` function is used to avoid issues.
 
 ## [1.3.0] 24-10-2024